[ArrowStringArray] CLN: assorted cleanup (#41306)
